[level 1] Title: x만큼 간격이 있는 n개의 숫자, Time: 0.03 ms, Memory: 78.4 MB -BaekjoonHub

+++ 프로그래머스/lv1/12954. x만큼 간격이 있는 n개의 숫자/x만큼 간격이 있는 n개의 숫자.java
@@ -1,13 +1,10 @@
-import java.math.BigInteger;
-
 class Solution {
-    public BigInteger[] solution(int x, int n) {
-        BigInteger[] answer = new BigInteger[n];
-
-        for (int i = 0; i < answer.length; i++) {
-            answer[i] = BigInteger.valueOf((long)x*(i+1));
+    public long[] solution(int x, int n) {
+        long[] answer = new long[n];
+        
+        for(int i = 0; i < answer.length; i++){
+            answer[i] = (long)x*(i+1);
         }
-
         return answer;
     }
 }
